% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textreport-package.R
\docType{package}
\name{textreport}
\alias{package-textreport}
\alias{textreport}
\alias{textreport-package}
\title{Auto Generate Descriptive Text Reports}
\description{
Automatically generate descriptive reports for text data.
}

