% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Title}
\usage{
report(data, grouping.var = NULL, tables.only = FALSE, open = TRUE,
  path = "textreport", text.var = TRUE, ...)
}
\arguments{
\item{data}{A data set with text to analyze.}

\item{grouping.var}{A character vector naming the grouping variable(s).}

\item{tables.only}{logical.  If \code{TRUE} the report will not be visual,
but a tables only version of the descriptive text analysis.}

\item{open}{logical.  If \code{TRUE} the \file{textreport.html} file will
attempt to be opened in the browser.}

\item{path}{A path to generate the files in.}

\item{text.var}{A character vector naming the text variable.  If \code{TRUE}
\code{report} will attempt to detect the text variable.}

\item{\ldots}{ignored.}
}
\description{
Description
}
\examples{
\dontrun{
report(presidential_debates_2012, c("person", "time"))
## Just Tables (non-viz report)
report(presidential_debates_2012, c("person", "time"), tables.only = TRUE)
}
}
\keyword{report}
\keyword{text}

